requirements.txt
requests~=2.31.0
tqdm~=4.65.0
python-dotenv~=1.0.0